refactor: Service, History, Test 수정
